refactor: colour radio set
